refactor: Restructure package
