+++ app/lib/screens/study/dashboard/contact_tab/contact_screen.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:provider/provider.dart';
-import 'package:studyou_core/models/config.dart';
-import 'package:studyou_core/models/study/contact.dart';
+import 'package:studyou_core/models/contact.dart';
+import 'package:studyou_core/models/models.dart';
 import 'package:studyou_core/util/retry_future_builder.dart';
 import 'package:url_launcher/url_launcher.dart';
