+++ app/lib/screens/study/onboarding/consent.dart
@@ -6,6 +6,7 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:pdf/widgets.dart' as pw;
 import 'package:provider/provider.dart';
+import 'package:studyou_core/models/consent/consent_item.dart';
 import 'package:studyou_core/models/models.dart';
 
 import '../../../models/app_state.dart';
