+++ app/lib/screens/study/onboarding/study_overview.dart
@@ -3,7 +3,6 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:provider/provider.dart';
 import 'package:studyou_core/models/models.dart';
-
 import 'package:studyou_core/util/supabase_future_builder.dart';
 
 import '../../../models/app_state.dart';
