+++ app/lib/screens/study/report/generic_section.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:studyou_core/models/study/studies.dart';
+import 'package:studyou_core/models/models.dart';
 
 abstract class GenericSection extends StatelessWidget {
   final UserStudy study;
