+++ app/lib/screens/study/report/sections/average_section_widget.dart
@@ -1,9 +1,8 @@
 import 'package:charts_flutter/flutter.dart' as charts;
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:studyou_core/models/report/sections/average_section.dart';
+import 'package:studyou_core/models/models.dart';
 import 'package:studyou_core/models/report/temporal_aggregation.dart';
-import 'package:studyou_core/models/study/studies.dart';
 
 import '../../../../util/data_processing.dart';
 import '../report_section_widget.dart';
