+++ core/lib/models/models.dart
@@ -1,5 +1,3 @@
-export 'config.dart';
-export 'consent/consent_item.dart';
 export 'data/data_reference.dart';
 export 'eligibility/eligibility_criterion.dart';
 export 'expressions/expression.dart';
@@ -8,7 +6,9 @@ export 'observations/observations.dart';
 export 'questionnaire/questionnaire_models.dart';
 export 'report/report_models.dart';
 export 'results/result.dart';
-export 'study/studies.dart';
 export 'study_results/study_results.dart';
 export 'study_schedule/study_schedule.dart';
+export 'tables/config.dart';
+export 'tables/study.dart';
+export 'tables/user_study.dart';
 export 'tasks/tasks.dart';
