--- core/lib/models/study/studies.dart
@@ -1,2 +1,0 @@
-export 'study.dart';
-export 'user_study.dart';
